microservices/auth/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
jwt
bcrypt
python-codicefiscale